Dim/Pos != null WIP
