+++ UnitTests/View/Layout/SetRelativeLayoutTests.cs
@@ -10,135 +10,7 @@ public class SetRelativeLayoutTests {
 	readonly ITestOutputHelper _output;
 
 	public SetRelativeLayoutTests (ITestOutputHelper output) => _output = output;
-
-	[Fact]
-	public void Null_Pos_Is_Same_As_PosAbsolute0 ()
-	{
-		var view = new View () {
-			X = null,
-			Y = null,
-		};
-
-		// Default layout style is Computed
-		Assert.Equal (LayoutStyle.Computed, view.LayoutStyle);
-		Assert.Null (view.X);
-		Assert.Null (view.Y);
-
-		view.BeginInit(); view.EndInit();
-
-		Assert.Equal (LayoutStyle.Computed, view.LayoutStyle);
-		Assert.Null (view.X);
-		Assert.Null (view.Y);
-
-		view.SetRelativeLayout (new Rect (5, 5, 10, 10));
-		Assert.Equal (LayoutStyle.Computed, view.LayoutStyle);
-		Assert.Null (view.X);
-		Assert.Null (view.Y);
-
-		Assert.Equal (0, view.Frame.X);
-		Assert.Equal (0, view.Frame.Y);
-	}
-
-	[Theory]
-	[InlineData (1, 1)]
-	[InlineData (0, 0)]
-	public void NonNull_Pos (int pos, int expectedPos)
-	{
-		var view = new View () {
-			X = pos,
-			Y = pos,
-		};
-
-		// Default layout style is Computed
-		Assert.Equal (LayoutStyle.Computed, view.LayoutStyle);
-		Assert.NotNull (view.X);
-		Assert.NotNull (view.Y);
-
-		view.BeginInit (); view.EndInit ();
-
-		Assert.Equal (LayoutStyle.Computed, view.LayoutStyle);
-		Assert.NotNull (view.X);
-		Assert.NotNull (view.Y);
-
-		view.SetRelativeLayout (new Rect (5, 5, 10, 10));
-		Assert.Equal (LayoutStyle.Computed, view.LayoutStyle);
-		Assert.NotNull (view.X);
-		Assert.NotNull (view.Y);
-
-		Assert.Equal (expectedPos, view.Frame.X);
-		Assert.Equal (expectedPos, view.Frame.Y);
-	}
-
-	[Fact]
-	public void Null_Dim_Is_Same_As_DimFill0 ()
-	{
-		var view = new View () {
-			Width = null,
-			Height = null,
-		};
-
-		// Default layout style is Computed
-		Assert.Equal (LayoutStyle.Computed, view.LayoutStyle);
-		Assert.Null (view.Width);
-		Assert.Null (view.Height);
-		view.BeginInit (); view.EndInit ();
-
-		Assert.Equal (LayoutStyle.Computed, view.LayoutStyle);
-		Assert.Null (view.Width);
-		Assert.Null (view.Height);
-
-		view.SetRelativeLayout (new Rect (5, 5, 10, 10));
-		Assert.Equal (LayoutStyle.Computed, view.LayoutStyle);
-		Assert.Null (view.Width);
-		Assert.Null (view.Height);
-		
-		Assert.Equal (0, view.Frame.X);
-		Assert.Equal (0, view.Frame.Y);
-
-		Assert.Equal (10, view.Frame.Width);
-		Assert.Equal (10, view.Frame.Height);
-
-		view.Width = Dim.Fill (0);
-		view.Height = Dim.Fill (0);
-		view.SetRelativeLayout (new Rect (5, 5, 10, 10));
-		Assert.Equal (10, view.Frame.Width);
-		Assert.Equal (10, view.Frame.Height);
-
-	}
-
-
-	[Theory]
-	[InlineData(1, 1)]
-	[InlineData (0, 0)]
-	public void NonNull_Dim (int dim, int expectedDim)
-	{
-		var view = new View () {
-			Width = dim,
-			Height = dim,
-		};
-
-		// Default layout style is Computed
-		Assert.Equal (LayoutStyle.Computed, view.LayoutStyle);
-		Assert.NotNull (view.Width);
-		Assert.NotNull (view.Height);
-		view.BeginInit (); view.EndInit ();
-
-		Assert.Equal (LayoutStyle.Computed, view.LayoutStyle);
-		Assert.NotNull (view.Width);
-		Assert.NotNull (view.Height);
-
-		view.SetRelativeLayout (new Rect (5, 5, 10, 10));
-		Assert.Equal (LayoutStyle.Computed, view.LayoutStyle);
-		Assert.NotNull (view.Width);
-		Assert.NotNull (view.Height);
-		
-		Assert.Equal (0, view.Frame.X);
-		Assert.Equal (0, view.Frame.Y);
-		// BUGBUG: Width == null is same as Dim.Absolute (0) (or should be). Thus this is a bug.
-		Assert.Equal (expectedDim, view.Frame.Width);
-		Assert.Equal (expectedDim, view.Frame.Height);
-	}
-
+	
 	[Fact]
 	public void Fill_Pos_Within_Bounds ()
 	{
@@ -263,7 +135,7 @@ public class SetRelativeLayoutTests {
 	}
 
 	[Fact]
-	public void FIll_And_PosCenter ()
+	public void Fill_And_PosCenter ()
 	{
 		var screen = new Rect (0, 0, 80, 25);
 		var view = new View () {
